Add buttons to delete all 3rd party metadata lookup info in the settings

+++ plexpy/datafactory.py
@@ -1406,16 +1406,29 @@ class DataFactory(object):
 
         return lookup_info
 
-    def delete_lookup_info(self, rating_key='', title=''):
+    def delete_lookup_info(self, rating_key='', service='', delete_all=False):
+        if not rating_key and not delete_all:
+            logger.error(u"Tautulli DataFactory :: Unable to delete lookup info: rating_key not provided.")
+            return False
+
         monitor_db = database.MonitorDatabase()
 
         if rating_key:
-            logger.info(u"Tautulli DataFactory :: Deleting lookup info for '%s' (rating_key %s) from the database."
+            logger.info(u"Tautulli DataFactory :: Deleting lookup info for rating_key %s from the database."
                         % (title, rating_key))
-            result_tvmaze = monitor_db.action('DELETE FROM tvmaze_lookup WHERE rating_key = ?', [rating_key])
             result_themoviedb = monitor_db.action('DELETE FROM themoviedb_lookup WHERE rating_key = ?', [rating_key])
+            result_tvmaze = monitor_db.action('DELETE FROM tvmaze_lookup WHERE rating_key = ?', [rating_key])
             result_musicbrainz = monitor_db.action('DELETE FROM musicbrainz_lookup WHERE rating_key = ?', [rating_key])
-            return True if (result_tvmaze or result_themoviedb or result_musicbrainz) else False
+            return bool(result_themoviedb or result_tvmaze or result_musicbrainz)
+        elif service and delete_all:
+            if service.lower() in ('themoviedb', 'tvmaze', 'musicbrainz'):
+                logger.info(u"Tautulli DataFactory :: Deleting all lookup info for '%s' from the database."
+                            % service)
+                result = monitor_db.action('DELETE FROM %s_lookup' % service.lower())
+                return bool(result)
+            else:
+                logger.error(u"Tautulli DataFactory :: Unable to delete lookup info: invalid service '%s' provided."
+                             % service)
 
     def get_search_query(self, rating_key=''):
         monitor_db = database.MonitorDatabase()
